q4 does NOT work
I am only uploading it because I am giving up on the question for the night but if you have a random burst of energy and want to attempt it while im sleeping, be my guest. But feel free to delete this all and start from scratch if that is easier.